+++ lib/app/config.dart
@@ -1,0 +1,5 @@
+class AppConfig {
+  AppConfig._();
+
+  static final launchDateTime = DateTime(2025, 1);
+}