+++ lib/ui/color.dart
@@ -8,6 +8,10 @@ extension AppColorScheme on ColorScheme {
     return Color(0xff318ad1);
   }
 
+  Color get onWater {
+    return Colors.white;
+  }
+
   Color get deepWater {
     if (brightness == Brightness.light) {
       return Color(0xff0665de);
@@ -15,10 +19,18 @@ extension AppColorScheme on ColorScheme {
     return Color(0xff1a6bd2);
   }
 
+  Color get onDeepWater {
+    return Colors.white;
+  }
+
   Color get warning {
     if (brightness == Brightness.light) {
       return Color(0xffc67b1a);
     }
     return Color(0xffcc8b38);
+  }
+
+  Color get onWarning {
+    return Colors.white;
   }
 }